% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBShiny2.R
\name{NBShiny2}
\alias{NBShiny2}
\title{Launch 'NBShiny2' Interface}
\usage{
NBShiny2()
}
\value{
Nothing
}
\description{
NBShiny2() loads interactive user interface built using R 'shiny'.
}
\details{
Start NBShiny2

The interactive user interface is to provide an easy way for naive bayesian classification analysis and downloading relevant plot.
}
\examples{
if(interactive()){
library(rmarkdown)
NBShiny2()
}
}
\keyword{NBShiny2}
